Phase 8: docs, repo hygiene, demo seed data, and test/migration fixes
- Rewrite README.md and add docs/architecture.md, ml-model-card.md,
  demo-script.md, demo-credentials.md, known-issues.md
- Add LICENSE (MIT), CONTRIBUTING.md, and tighten .gitignore
- Add idempotent demo seeders (Faculty/User/Subject/Student/Attendance)
  producing 9 users, 200 students, 40 subjects, ~52k attendance rows
- Fix FK type/signedness mismatches (errno 3780) in email_logs,
  detention_predictions, and hod_attendance_summary migrations, and
  fix information_schema column casing in the partition migration
- Add Laravel feature/unit tests and ML service pytest suite (55 passed,
  82% coverage)

+++ ATTENDANCE/attendance-email-system/api/database/migrations/2026_06_07_162504_create_email_logs_table.php
@@ -10,7 +10,10 @@ return new class extends Migration
     {
         Schema::create('email_logs', function (Blueprint $table) {
             $table->id('log_id');
-            $table->unsignedBigInteger('student_id');
+            // students.student_id is a signed INT (legacy schema_mysql.sql),
+            // so the FK column must match exactly (type + signedness) or
+            // MySQL rejects the foreign key with errno 3780.
+            $table->integer('student_id');
             $table->string('recipient_email');
             $table->enum('email_type', ['daily_attendance', 'detention_notice', 'monthly_report']);
             $table->enum('status', ['queued', 'sent', 'failed'])->default('queued');
